+++ actions/agents-bootstrapper/actions.py
@@ -8,7 +8,7 @@ from urllib.parse import urljoin
 import black
 import requests
 from refresh_agent_spec_helper import update_agent_spec
-from sema4ai.actions import Response, action
+from sema4ai.actions import ActionError, Response, action
 
 AGENTS_DIR_PATH = Path.home() / "agents_bootstrapper"
 ACTIONS_GITHUB_URL = "https://api.github.com/repos/Sema4AI/gallery/contents/actions"
@@ -38,7 +38,7 @@ def bootstrap_agent_package(agent_name: str) -> Response[str]:
 
 
 @action
-def open_agent_code(agent_name: str) -> str:
+def open_agent_code(agent_name: str) -> Response[str]:
     """
     This action opens the code of the agent package with VSCode.
 
@@ -52,16 +52,20 @@ def open_agent_code(agent_name: str) -> str:
     full_agent_path = AGENTS_DIR_PATH / agent_name
 
     if not os.path.exists(full_agent_path):
-        return f"Error: agent package '{agent_name}' does not exist at path {full_agent_path}."
+        raise ActionError(
+            "Error: agent package '{agent_name}' does not exist at path {full_agent_path}."
+        )
 
     command = ["code", str(full_agent_path)]
 
     try:
         subprocess.run(command, check=True)
     except Exception as e:
-        return f"Unexpected error: {str(e)}. " "Please check your setup and try again."
+        raise ActionError(
+            f"Unexpected error: {str(e)}. " "Please check your setup and try again."
+        )
 
-    return f"{agent_name} code opened with VSCode."
+    return Response(result=f"{agent_name} code opened with VSCode.")
 
 
 @action
@@ -76,7 +80,7 @@ def refresh_agent_package_spec(agent_name: str) -> Response[None]:
 
 
 @action
-def list_available_prebuilt_actions() -> list[str]:
+def list_available_prebuilt_actions() -> Response[list[str]]:
     """
     List all folders (actions) inside the 'actions' directory of the repository.
 
@@ -87,13 +91,13 @@ def list_available_prebuilt_actions() -> list[str]:
     if response.status_code == 200:
         data = response.json()
         folders = [item["name"] for item in data if item["type"] == "dir"]
-        return folders
+        return Response(result=folders)
     else:
-        return []
+        return Response(result=[])
 
 
 @action
-def read_prebuild_action_capabilities(action_name: str) -> str:
+def read_prebuild_action_capabilities(action_name: str) -> Response[str]:
     """
     Read the capabilities of a prebuild action package that is available in Github.
 
@@ -111,11 +115,15 @@ def read_prebuild_action_capabilities(action_name: str) -> str:
             import base64
 
             readme_content = base64.b64decode(data["content"]).decode("utf-8")
-            return readme_content
+            return Response(result=readme_content)
         else:
-            return "README.md file is empty or not available in the expected format."
+            raise ActionError(
+                "README.md file is empty or not available in the expected format."
+            )
     else:
-        return f"Unable to fetch README.md. Status Code: {response.status_code}, error: {response.text}"
+        raise ActionError(
+            f"Unable to fetch README.md. Status Code: {response.status_code}, error: {response.text}"
+        )
 
 
 def download_file(url, save_path):
@@ -131,7 +139,9 @@ def download_file(url, save_path):
         with open(save_path, "wb") as file:
             file.write(response.content)
     else:
-        print(f"Failed to download file: {url}. Status Code: {response.status_code}")
+        raise ActionError(
+            f"Failed to download file: {url}. Status Code: {response.status_code}"
+        )
 
 
 def download_folder(url: str, local_path: Path) -> str:
@@ -154,13 +164,15 @@ def download_folder(url: str, local_path: Path) -> str:
             elif item["type"] == "dir":
                 download_folder(item["url"], local_path / item["name"])
     else:
-        return f"Failed to download action: {url}. Status Code: {response.status_code}, message: {response.text}"
+        raise ActionError(
+            f"Failed to download action: {url}. Status Code: {response.status_code}, message: {response.text}"
+        )
 
     return "Action downloaded successfully."
 
 
 @action
-def download_prebuilt_action(action_name: str, agent_name: str) -> str:
+def download_prebuilt_action(action_name: str, agent_name: str) -> Response[str]:
     """
     Downloads a prebuilt action package from Github to the specified agent package.
     This method requires an agent to be bootstrapped first.
@@ -174,14 +186,14 @@ def download_prebuilt_action(action_name: str, agent_name: str) -> str:
     """
     agent_path = AGENTS_DIR_PATH / agent_name
     if not agent_path.exists():
-        return (
+        raise ActionError(
             f"Error: agent package '{agent_name}' does not exist at path {agent_path}."
         )
 
     url = f"{ACTIONS_GITHUB_URL}/{action_name}"
     sema4_actions_path = agent_path / "actions" / SEMA4AI_ACTIONS / action_name
 
-    return download_folder(url, sema4_actions_path)
+    return Response(result=download_folder(url, sema4_actions_path))
 
 
 def get_sema4_ai_studio_url_for_agent_zip_path(path: str) -> str: